+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-16T12:13:19.225Z",
+    "2026-05-17T12:53:05.298Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {