+++ packages/orq-rc/RELEASES.md
@@ -9239,3 +9239,13 @@ Based on:
 - [typescript v4.10.0-rc.9] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.9 - packages/orq-rc
+
+## 2026-05-17 12:56:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.1 (2.884.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.10] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.10] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.10 - packages/orq-rc